// src/api/generated/user-order-service.schemas.ts
export type OrderSummaryResponseDtoDeliveryStatus = typeof OrderSummaryResponseDtoDeliveryStatus[keyof typeof OrderSummaryResponseDtoDeliveryStatus];


export const OrderSummaryResponseDtoDeliveryStatus = {
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  COMPLETED: 'COMPLETED',
} as const;

export type OrderSummaryResponseDtoPaymentStatus = typeof OrderSummaryResponseDtoPaymentStatus[keyof typeof OrderSummaryResponseDtoPaymentStatus];


export const OrderSummaryResponseDtoPaymentStatus = {
  INITIAL: 'INITIAL',
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
} as const;